+++ demos/VertexAnimator.Sinoid.js
@@ -1,5 +1,5 @@
 /**
- * @classdesc Animate a set of vertices: let them move in a cicular manner. Multiple revolutions possible.
+ * @classdesc Animate a set of vertices: let them move in a cicular-sinoidal manner.
  *
  * Note that this class not not very robust towards to vertex array modifications. Please
  * create a new animator once your vertex list changed.
@@ -7,10 +7,10 @@
  * @requires Vertex
  *
  * @author Ikaros Kappler
- * @date 2019-10-30 (Cloned from LinearVertexAnimator).
+ * @date 2020-08-19 (Cloned from CircularVertexAnimator).
  * @version 1.0.0
  *
- * @file VertexAnimator.Circular.js
+ * @file VertexAnimator.Sinoid.js
  * @public
  **/
 
@@ -33,22 +33,34 @@
 	var s = Math.sin, c = Math.cos, PI = Math.PI, TWOPI = Math.PI*2;
 	let angle = 0, radius = 0;
 	var maxRadius = Math.min( viewport.max.x-viewport.min.x, viewport.max.y-viewport.min.y ) * 0.45;
-	var minRadius = maxRadius*0.77;
+	var minRadius = maxRadius*0.80;
 	var startTime = 0;
-	var revolutions = 3;
+	var revolutions = 1;
+
+	var vertexTimeOffsets = [];
+	var vertexRadiusFactors = [];
+	for( var i = 0; i < vertices.length; i++ ) {
+	    vertexTimeOffsets.push( Math.random() );
+	    vertexRadiusFactors.push( Math.random() );
+	}
 	
 	function animateVert( vert, index, count, time, viewport ) {
 	    if( !startTime ) startTime = time;
 	    time -= startTime;
+
 	    
-	    angle = TWOPI*(index/count) + time/20000;
-	    radius = -1.0-s(angle+time/1000); // +c(angle+time/100000);
-	    radius = minRadius + (maxRadius-minRadius)*radius;
+	    angle = TWOPI*(index/count) + time/10000;
+	    radius = -1.0-s(angle+time/1000 + vertexTimeOffsets[index]*10); // +c(angle+time/100000);
+	    radius = minRadius + (maxRadius-minRadius)*radius * vertexRadiusFactors[index];
 	    // Add some noise
-	    radius += s((angle*20))*11;
+	    // radius += s((angle*20))*11;
 	    // Apply angle and radius and noise
 	    vert.x = c(angle*revolutions)*radius;
 	    vert.y = s(angle*revolutions)*radius;
+
+	    
+
+	    
 	};
 
 	function animateAll( time ) {
@@ -85,6 +97,6 @@
 	};
     };
 
-    _context.CircularVertexAnimator = VertexAnimator;
+    _context.SinoidVertexAnimator = VertexAnimator;
 
 })(window ? window : module.export);